Move away from VIPS for bimg
We noticed that our images are being rotated in sherlock.

It seems that this is happening in gothumb (we have the original in the correct orientation)

Rather than try and work this out I found this issue. https://github.com/DAddYE/vips/issues/37

What is concerning is that from the README

> When generating JPEG output all metadata is removed and Huffman tables optimised without having to use separate command line tools like jpegoptim and jpegtran.

I'm guessing at Go a bit so there's likely some unwanted changes in this :(